move jexl to scm-core
