+++ pom.xml
@@ -32,7 +32,7 @@
   <parent>
     <artifactId>scm-plugins</artifactId>
     <groupId>sonia.scm.plugins</groupId>
-    <version>2.0.0</version>
+    <version>2.3.0-SNAPSHOT</version>
   </parent>
 
   <artifactId>scm-webhook-plugin</artifactId>
@@ -45,17 +45,10 @@
   <dependencies>
 
     <dependency>
-      <groupId>org.apache.commons</groupId>
-      <artifactId>commons-jexl</artifactId>
-      <version>2.1.1</version>
-    </dependency>
-
-    <dependency>
       <groupId>org.codehaus.jackson</groupId>
       <artifactId>jackson-mapper-asl</artifactId>
       <version>1.9.2</version>
     </dependency>
-
 
     <!-- test scope -->
